v2/requirements.txt
python-telegram-bot
nltk
pymongo
matplotlib
python-dotenv
torch 
torchvision 
torchaudio
googletrans-py
laser_encoders